Extract certificate compression tests
